+++ include/keyboard.hpp
@@ -1,5 +1,6 @@
 #pragma once
 #include <common.hpp>
+#include <math.hpp>
 
 
 namespace rack {
@@ -13,6 +14,8 @@ namespace keyboard {
 void init();
 void press(int key);
 void release(int key);
+/** pos is in the unit box. */
+void mouseMove(math::Vec pos);
 
 
 } // namespace keyboard